feat: update file size limits to 1000 lines across entire project
## File Size Limits Update (500 → 1000 lines)

### Agent Rules Updated:
- Updated all agent documentation (.mdc files) to use 1000 lines limit
- Updated API Designer, Backend, Content Writer, Database, DevOps, etc.
- Updated performance, security, QA, UI/UX, UE5, Release agents

### Infrastructure Scripts Updated:
- Updated all Makefile files in services with 1000 lines checks
- Updated validation scripts (architecture, file-structure)
- Updated git hooks and pre-commit checks

### Validation Tools Updated:
- Updated architecture-validator.py with bundled file exclusions
- Updated check-all-files.sh with comprehensive exclusions
- Updated all PowerShell and Bash validation scripts

### Bundled Files Exclusions:
- Added docker-compose.yml to all exclusions
- Added comprehensive bundled.yaml patterns (*-bundled.yaml)
- Added generated file patterns (_gen.go, .pb.go, oas_*.go)
- Updated pre-commit-exemptions.txt with all patterns

### Documentation Updated:
- Updated README-validation.md with correct examples
- Updated agent guides and file placement rules
- Updated performance optimization checklists

All text files now have consistent 1000-line limit with proper exclusions
for generated/bundled files and infrastructure configurations.

+++ services/realtime-gateway-go/server/game_state.go
@@ -53,7 +53,7 @@ func (gsm *GameStateManager) UpdatePlayerInput(input *PlayerInputData) {
 	}
 
 	speed := float32(5.0)
-	dt := float32(1.0 / float32(gsm.tickRate))
+	dt := 1.0 / float32(gsm.tickRate)
 
 	moveX := DequantizeCoordinate(input.MoveX)
 	moveY := DequantizeCoordinate(input.MoveY)
@@ -99,4 +99,3 @@ func (gsm *GameStateManager) RemovePlayer(playerID string) {
 	defer gsm.mu.Unlock()
 	delete(gsm.players, playerID)
 }
-
